DjangoDir/PRDProject/requirements.txt
asgiref==3.2.10
dj-database-url==0.5.0
Django==3.1.2
django-cors-headers==3.6.0
django-crispy-forms==1.10.0
django-heroku==0.3.1
djangorestframework==3.12.2
Pillow==8.1.0
psycopg2==2.8.6
python-decouple==3.4
pytz==2020.1
sqlparse==0.4.1
whitenoise==5.2.0
